+++ src/ClassDiagramBuilderInterface.php
@@ -10,23 +10,23 @@ interface ClassDiagramBuilderInterface
 {
     public const OPTIONS_DEFAULTS = [
         // string to indent graph statement parts
-        'indent-string' => '  ',
+        'indent_string' => '  ',
         // whether to use html or record formatted labels (graphviz specific feature)
-        'label-format' => 'record',
+        'label_format' => 'record',
         // whether to only show methods/properties that are actually defined in this class (and not those merely inherited from base)
-        'only-self' => true,
+        'only_self' => true,
         // whether to also show private methods/properties
-        'show-private' => true,
+        'show_private' => true,
         // whether to also show protected methods/properties
-        'show-protected' => true,
+        'show_protected' => true,
         // whether to show class constants as readonly static variables (or just omit them completely)
-        'show-constants' => true,
+        'show_constants' => true,
         // whether to show class properties
-        'show-properties' => true,
+        'show_properties' => true,
         // whether to show class or interface methods
-        'show-methods' => true,
+        'show_methods' => true,
         // whether to show add parent classes or interfaces
-        'add-parents' => true,
+        'add_parents' => true,
     ];
 
     public function createVertexClass($class): Vertex;